Fixed linting Created a function to create a common rollup config
