+++ packages/jss/tests/.babelrc
@@ -1,5 +1,5 @@
 {
-  extends: '../.babelrc',
+  extends: '../../../.babelrc',
   env: {
     test: {
       plugins: ['transform-runtime']
